ecommerce/inventory/serializers.py
""" serializers to parse into json using in DRF """
from rest_framework import serializers
from . import models

class ProductSerializer(serializers.ModelSerializer):
    class Meta:
        model = models.Product
        fields = ['id','name','is_active']